requirements.txt
Django==4.2.3
django-environ==0.10.0
django-pagination==1.0.7
django-simple-pagination==1.3
djangorestframework==3.14.0
Pillow==10.0.0
psycopg2==2.9.6
python-dotenv==1.0.0
sqlparse==0.4.4
django-crispy-forms